+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## X.Y.Z (Unreleased)
 
-...
+ENHANCEMENTS:
+
+* resource/sonatypeiq_organization: Now supports `terraform import` [GH-46]
+
 
 ## 0.10.0 July 14, 2025